network: EMAC/PHY and core fixes
Multiple fixes and refactors across lib-network: switch PHY API to use uint16_t addresses and add Unknown enum values for Duplex/Speed; expand MMI advertise mask to include CSMA; unify and rename GD32 PTP/time functions (e.g. gd32::PtpSubsecond2Nanosecond, Gd32PtpGetTime, Gd32PtpUpdateTime, Gd32PtpStart); replace hal::Millis()/hal timing with timing::Millis(); centralize error reporting via network::Error/ERROR and remove console::Error declarations; update network memory allocator to call network::Error; adjust EMAC counters to 32-bit; tighten PHY parsing and autonegotiation logic (advertisement/config/restart, link parsing, ToString helpers) and improve debug prints; rename DP83848 PHY source files to emac_* names; small API/variable renames and cleanup (kPhyValue, kAddress, IWYU pragmas). Also minor fixes in DHCP/IGMP/UDP/NTP code to use new timing/PTP APIs and improved error messages.

+++ lib-network/src/emac/gd32/emac_counters.cpp
@@ -30,8 +30,8 @@
 #include "gd32.h" // IWYU pragma: keep
 
 namespace network::iface {
-static uint64_t s_rx_fifo_drop_total = 0;
-static uint64_t s_rx_dma_missed_total = 0;
+static uint32_t s_rx_fifo_drop_total = 0;
+static uint32_t s_rx_dma_missed_total = 0;
 
 static void PollHardwareCounters() {
     uint32_t rxfifo_drop;  ///< The number of frames dropped by RxFIFO
